extract LemmaMatcher and SignMatcher
